+++ packages/eyes-storybook/test/e2e/fake-ie.e2e.test.js
@@ -8,7 +8,7 @@ const {version} = require('../../package.json');
 const envWithColor = {...process.env, FORCE_COLOR: true};
 const spawnOptions = {stdio: 'pipe', env: envWithColor};
 
-describe.only('fake ie', () => {
+describe('fake ie', () => {
   it('fake ie in storybook', async () => {
     const [err, result] = await presult(
       utils.process.sh(